requirements.txt
pandas>=2.0.0
pyodbc>=4.0.39
xlsxwriter>=3.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0

# Database
sqlalchemy>=2.0.0

# Development dependencies (optional)
black>=23.0.0        # Code formatter
flake8>=6.0.0        # Linter

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1